flask-backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0